+++ comment/printer.go
@@ -11,7 +11,7 @@ import (
 
 func ToString(comments endpoints.Comments, indentSize int, commentWidth int, screenWidth int, preserveRightMargin bool,
 	altIndentBlock bool, commentHighlighting bool) string {
-	header := getHeader(comments, commentWidth, screenWidth, commentHighlighting)
+	header := getHeader(comments, commentWidth, screenWidth, commentHighlighting, altIndentBlock)
 	replies := ""
 
 	for _, reply := range comments.Comments {
@@ -22,14 +22,15 @@ func ToString(comments endpoints.Comments, indentSize int, commentWidth int, scr
 	return header + replies
 }
 
-func getHeader(c endpoints.Comments, commentWidth int, screenWidth int, commentHighlighting bool) string {
+func getHeader(c endpoints.Comments, commentWidth int, screenWidth int, commentHighlighting bool,
+	altIndentBlock bool) string {
 	if commentWidth == 0 {
 		commentWidth = screenWidth
 	}
 
 	headline := getHeadline(c.Title, c.Domain, c.URL, c.ID, commentWidth)
 	infoLine := getInfoLine(c.Points, c.User, c.TimeAgo, c.CommentsCount, c.ID)
-	rootComment := parseRootComment(c.Content, commentWidth, commentHighlighting)
+	rootComment := parseRootComment(c.Content, commentWidth, commentHighlighting, altIndentBlock)
 	helpMessage := dimmed(messages.LessScreenInfo) + NewLine
 	separator := messages.GetSeparator(commentWidth)
 
@@ -73,12 +74,12 @@ func getHyperlinkText(url string, text string) string {
 	return Link1 + url + Link2 + text + Link3
 }
 
-func parseRootComment(c string, commentWidth int, commentHighlighting bool) string {
+func parseRootComment(c string, commentWidth int, commentHighlighting bool, altIndentBlock bool) string {
 	if c == "" {
 		return ""
 	}
 
-	comment, URLs := ParseComment(c, commentWidth, commentWidth, commentHighlighting)
+	comment, URLs := ParseComment(c, commentWidth, commentWidth, commentHighlighting, altIndentBlock)
 	wrappedComment, _ := text.Wrap(comment, commentWidth)
 	wrappedComment = applyURLs(wrappedComment, URLs)
 
@@ -89,7 +90,7 @@ func printReplies(c endpoints.Comments, indentSize int, commentWidth int, screen
 	parentPoster string, preserveRightMargin bool, altIndentBlock bool, commentHighlighting bool) string {
 	currentIndentSize := indentSize * c.Level
 	usableScreenSize := screenWidth - currentIndentSize
-	comment, URLs := ParseComment(c.Content, commentWidth, usableScreenSize, commentHighlighting)
+	comment, URLs := ParseComment(c.Content, commentWidth, usableScreenSize, commentHighlighting, altIndentBlock)
 	adjustedCommentWidth := getCommentWidthForLevel(c.Level, indentSize, commentWidth, screenWidth, preserveRightMargin)
 
 	indentBlock := getIndentBlock(c.Level, indentSize, altIndentBlock)